Ajuste de altura e largura, número páginas e quantidade

+++ SME.CDEP.Aplicacao/DTOS/AcervoFotograficoCadastroDTO.cs
@@ -19,8 +19,8 @@ public class AcervoFotograficoCadastroDTO : AcervoCadastroDTO
     public long ConservacaoId { get; set; }
     
     [Required(ErrorMessage = "É necessário informar a quantidade do acervo fotografico")]
-    [Range(1, long.MaxValue, ErrorMessage = "A quantidade do acervo fotográfico deve ser maior que zero")]
-    public long Quantidade { get; set; }
+    [Range(1, int.MaxValue, ErrorMessage = "A quantidade do acervo fotográfico deve ser maior que zero")]
+    public int Quantidade { get; set; }
     
     public string? Largura { get; set; }
     public string? Altura { get; set; }
