+++ SME.CDEP.Aplicacao/DTOS/AcervoTridimensionalCadastroDTO.cs
@@ -13,8 +13,8 @@ public class AcervoTridimensionalCadastroDTO : AcervoCadastroDTO
     public long ConservacaoId { get; set; }
     
     [Required(ErrorMessage = "É necessário informar a quantidade do acervo arte gráfica")]
-    [Range(1, long.MaxValue, ErrorMessage = "A quantidade do acervo arte gráfica deve ser maior que zero")]
-    public long Quantidade { get; set; }
+    [Range(1, int.MaxValue, ErrorMessage = "A quantidade do acervo arte gráfica deve ser maior que zero")]
+    public int Quantidade { get; set; }
     public string? Largura { get; set; }
     public string? Altura { get; set; }
     public string? Profundidade { get; set; }
