+++ SME.CDEP.Dominio/Entidades/AcervoTridimensional.cs
@@ -6,7 +6,7 @@
         public long AcervoId { get; set; }
         public string Procedencia { get; set; }
         public long ConservacaoId { get; set; }
-        public long Quantidade { get; set; }
+        public int Quantidade { get; set; }
         public string? Largura { get; set; }
         public string? Altura { get; set; }
         public string? Profundidade { get; set; }
